migrated EventManager code
