+++ src/main/java/org/braekpo1nt/mctmanager/games/event/EventState.java
@@ -1,0 +1,10 @@
+package org.braekpo1nt.mctmanager.games.event;
+
+public enum EventState {
+    WAITING_HUB,
+    VOTING,
+    TELEPORTING_TO_GAME,
+    PLAYING_GAME,
+    TELEPORTING_FROM_GAME,
+    PAUSED,
+}
